requirements.txt
# TruPhoto — Python Dependencies
# Install: pip install -r requirements.txt
#
# Bounds are conservative: known-good major ranges that all support
# Python >= 3.9. Tighten to exact pins (use `pip freeze`) once the
# pipeline is validated end-to-end.

# Core ML
torch>=2.0,<3
torchvision>=0.15,<1
timm>=1.0,<2
scikit-learn>=1.3,<2
numpy>=1.24,<2
joblib>=1.3,<2

# Data handling
pandas>=2.0,<2.3
Pillow>=10,<12

# Visualization
matplotlib>=3.7,<4
seaborn>=0.13,<1

# Demo interface
gradio>=4,<6

# Utilities
tqdm>=4.65,<5
